feat: refatoração mvc tela projetos
