+++ src/App.js
@@ -5,7 +5,7 @@ import { UserDetailsProvider } from './context/usercontext';
 import Home from './pages/home/home.component';
 import { Navigate } from 'react-router-dom';
 import { ProjectDetailsProvider } from './context/projectContext';
-import Projetos from './pages/projetos/projetos.component';
+import ProjetosController from './pages/projetos/ProjetosController';
 import Dashboard from './pages/dashboard/dashboard.component';
 import Backlog from './pages/backlog/backlog.component';
 import Painel from './pages/painel/painel.component';
@@ -20,7 +20,7 @@ const App = () => {
           <Route path="/" index element={<Home />}></Route>
           <Route path="auth/" element={<AuthenticationController />} />
           <Route path="*" element={<Navigate to="/" />} />
-          <Route path="projects" element={<Projetos />} />
+          <Route path="projects" element={<ProjetosController />} />
           <Route path="projects/:projectId/dashboard" element={<Dashboard />} />
           <Route path="projects/:projectId/backlog" element={<Backlog />} />
           <Route
